Move public async_connect function into impl
Public interface to establish connection is connection_info
provider. So the async_connect is the implementation only
function.
