+++ include/ozo/connection_info.h
@@ -1,14 +1,14 @@
 #pragma once
 
 #include <ozo/connection.h>
-#include <ozo/async_connect.h>
+#include <ozo/impl/async_connect.h>
 
 #include <chrono>
 
 namespace ozo {
 
 template <
-    typename OidMap = empty_oid_map, 
+    typename OidMap = empty_oid_map,
     typename Statistics = no_statistics>
 class connection_info {
     io_context& io_;
@@ -26,7 +26,7 @@ public:
 
     template <typename Handler>
     friend void async_get_connection(const connection_info& self, Handler&& h) {
-        async_connect(self.conn_str_,
+        impl::async_connect(self.conn_str_,
             std::make_shared<connection>(self.io_, self.statistics_),
             std::forward<Handler>(h));
     }
